add rl_eof_found to the api, and set correctly for callback api
Currently _rl_eof_found is only set for the traditional (non-callback)
api.  This commit ensures that _rl_eof_found is also set for the
callback api.

Additionally, this commit renamed _rl_eof_found to rl_eof_found, and
makes this variable part of the advertised api.  The documentation is
updated to include mention of this variable.